requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
telethon==1.34.0
openpyxl==3.1.2
requests==2.31.0
